Shared interactive node and multinode MPI test

+++ Makefile
@@ -10,6 +10,6 @@ install: $(EXES) $(HELPERS) $(SCRIPTS)
 	install -o root -g root -m 0700 -t $(PREFIX)/bin $(SCRIPTS_ROOT)
 	install -o root -g root -m 0755 -t $(PREFIX)/bin $(SCRIPTS_USER)
 	install -o root -g root -m 0755 -d $(PREFIX)/etc $(PREFIX)/libexec/dgx-cgroup
-	install -o root -g root -m 0700 -t $(PREFIX)/etc etc/$(ETC)
+	if [ ! -f $(PREFIX)/etc/$(ETC) ] ; then install -o root -g root -m 0700 -t $(PREFIX)/etc etc/$(ETC) ; fi
 	for f in $(LIBEXEC) ; do install -o root -g root -m 0700 -t $(PREFIX)/libexec/dgx-cgroup libexec/$$f ; done
 	./pam-config
